chatbot app/cbrequirements.txt
openai>=1.0.0
python-dotenv
streamlit
